+++ Frontend/app/forms/dynamic-forms.module.ts
@@ -8,7 +8,8 @@ import {RouterModule, Routes} from '@angular/router';
 import {FormsService} from './services/forms.service';
 import {HAMMER_GESTURE_CONFIG, HammerGestureConfig} from '@angular/platform-browser';
 import {SharedModule} from '../shared/shared.module';
-import {TextareaComponent} from './components/textarea.component';
+import {TextareaUndoComponent} from './components/textarea-undo.component';
+import {SingleLineTextInputComponent} from './components/single-line-text-input.component';
 
 const ROUTES: Routes = [
 ];
@@ -17,7 +18,8 @@ const COMPONENTS = [
 	ListComponent,
 	InputComponent,
 	FormComponent,
-	TextareaComponent
+	TextareaUndoComponent,
+	SingleLineTextInputComponent
 ];
 
 const MODULES = [