+++ Frontend/app/main/components/companies/companies-details.component.ts
@@ -3,13 +3,13 @@ import {Component} from '@angular/core';
 	selector: 'company-details-component',
 	template: `
 		Company Details
-		<textarea-component label="Name"></textarea-component>
-		<textarea-component label="Address"></textarea-component>
-		<textarea-component label="City"></textarea-component>
-		<textarea-component label="Zip"></textarea-component>
-		<textarea-component label="Phone"></textarea-component>
-		<textarea-component label="Web"></textarea-component>
-		<textarea-component label="Misc"></textarea-component>
+		<single-line-text-input-component label="Name"></single-line-text-input-component>
+		<single-line-text-input-component label="Address"></single-line-text-input-component>
+		<single-line-text-input-component label="City"></single-line-text-input-component>
+		<single-line-text-input-component label="Zip"></single-line-text-input-component>
+		<single-line-text-input-component label="Phone"></single-line-text-input-component>
+		<single-line-text-input-component label="Web"></single-line-text-input-component>
+		<single-line-text-input-component label="Misc"></single-line-text-input-component>
 		<button type="button" class="btn-success btn-lg pull-right">Save</button>
 		<button type="button" class="btn-warning btn-lg pull-right">Cancel</button>
 	`