+++ Frontend/app/main/components/quotes/add-quote.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {Component} from '@angular/core';
 
 @Component({
 	selector: 'create-quote-component',
@@ -8,56 +8,74 @@ import {Component, OnInit} from '@angular/core';
 		<quote-header>
 			<div class="row">
 				<div class="col-xs-12 text-center">Quote Title</div>
-				<textarea-component></textarea-component>
+				<single-line-text-input-component></single-line-text-input-component>
 				<hr>
 			</div>
 		</quote-header>
 		<quote-body>
-			<div class="row">
-				<div class="row">
-					<quote-line-description class="col-xs-8">Quote Line Description</quote-line-description>
-					<span class="col-xs-4 icon icon-cross"></span>
-				</div>
-				<div class="row">
-					<quote-line-qty class="col-xs-4">42</quote-line-qty>
-					<quote-line-price class="col-xs-8">$163.00</quote-line-price>
-				</div>
-			</div>
-
-			<div class="row">
-				<div class="row">
-					<quote-line-description class="col-xs-8">Quote Line Description</quote-line-description>
-					<span class="col-xs-4 icon icon-cross"></span>
-				</div>
-				<div class="row">
-					<quote-line-qty class="col-xs-4">42</quote-line-qty>
-					<quote-line-price class="col-xs-8">$163.00</quote-line-price>
-				</div>
-			</div>
-
-			<div class="row">
-				<div class="row">
-					<quote-line-description class="col-xs-8">Quote Line Description</quote-line-description>
-					<span class="col-xs-4 icon icon-cross"></span>
-				</div>
-				<div class="row">
-					<quote-line-qty class="col-xs-4">42</quote-line-qty>
-					<quote-line-price class="col-xs-8">$163.00</quote-line-price>
-				</div>
-			</div>
+			<ul class="quote-list">
+				<li class="quote-list-line">
+					<span class="icon icon-cross pull-right"></span>
+					<div class="quote-list-line-title-wrapper">
+						<strong>Quote Line Description</strong>
+					</div>
+					<div class="quote-list-line-details">
+						QTY
+						<single-line-text-input-component [model]="'42'"></single-line-text-input-component>
+						$
+						<single-line-text-input-component [model]="'163.00'"></single-line-text-input-component>
+						<quote-line-options>
+							<span class="icon icon-arrow-up"></span>
+							<span class="icon icon-arrow-down"></span>
+						</quote-line-options>
+					</div>
+				</li>
+				<li class="quote-list-line">
+					<span class="icon icon-cross pull-right"></span>
+					<div class="quote-list-line-title-wrapper">
+						<strong>Quote Line Description</strong>
+					</div>
+					<div class="quote-list-line-details">
+						QTY
+						<single-line-text-input-component [model]="'42'"></single-line-text-input-component>
+						$
+						<single-line-text-input-component [model]="'163.00'"></single-line-text-input-component>
+						<quote-line-options>
+							<span class="icon icon-arrow-up"></span>
+							<span class="icon icon-arrow-down"></span>
+						</quote-line-options>
+					</div>
+				</li>
+				<li class="quote-list-line">
+					<span class="icon icon-cross pull-right"></span>
+					<div class="quote-list-line-title-wrapper">
+						<strong>Quote Line Description</strong>
+					</div>
+					<div class="quote-list-line-details">
+						QTY
+						<single-line-text-input-component [model]="'42'"></single-line-text-input-component>
+						$
+						<single-line-text-input-component [model]="'163.00'"></single-line-text-input-component>
+						<quote-line-options>
+							<span class="icon icon-arrow-up"></span>
+							<span class="icon icon-arrow-down"></span>
+						</quote-line-options>
+					</div>
+				</li>
+			</ul>
 		</quote-body>
 		<quote-footer>
 			<div class="row">
 				<hr>
-				<textarea-component label="Desc."></textarea-component>
-				<textarea-component label="Unit Qty."></textarea-component>
-				<textarea-component label="Cost"></textarea-component>
+				<single-line-text-input-component label="Desc."></single-line-text-input-component>
+				<single-line-text-input-component label="Unit Qty."></single-line-text-input-component>
+				<single-line-text-input-component label="Cost"></single-line-text-input-component>
 				<button class="btn btn-lg">Add Line</button>
 			</div>
 		</quote-footer>
 		<hr>
-		<button class="btn-success btn-lg pull-right">Save Quote</button>
 		<button class="btn-warning btn-lg pull-right">Cancel</button>
+		<button class="btn-success btn-lg pull-right">Save</button>
 	`
 })
 export class AddQuoteComponent {}