+++ Frontend/app/main/components/user/user-settings.component.ts
@@ -2,15 +2,15 @@ import {Component} from '@angular/core';
 @Component({
 	selector: 'settings',
 	template: `
-	<h5>Your Settings:</h5>
-	<textarea-component label="Name"></textarea-component>
-	<textarea-component label="Email"></textarea-component>
-	<textarea-component label="Address"></textarea-component>
-	<textarea-component label="WebSite"></textarea-component>
-	<textarea-component label="Phone"></textarea-component>
-	<textarea-component label="Fax"></textarea-component>
-	<button class="btn-success btn-lg pull-right" [routerLink]="['/Home']">OK</button>
-	<button class="btn-warning btn-lg pull-right">Cancel</button>
+		<h5>Your Settings:</h5>
+		<single-line-text-input-component label="Name"></single-line-text-input-component>
+		<single-line-text-input-component label="Email"></single-line-text-input-component>
+		<single-line-text-input-component label="Address"></single-line-text-input-component>
+		<single-line-text-input-component label="WebSite"></single-line-text-input-component>
+		<single-line-text-input-component label="Phone"></single-line-text-input-component>
+		<single-line-text-input-component label="Fax"></single-line-text-input-component>
+		<button class="btn-warning btn-lg pull-right">Cancel</button>
+		<button class="btn-success btn-lg pull-right" [routerLink]="['/Home']">Save</button>
 	`
 })
 export class UserSettingsComponent {}